+++ packages/shared/sdk-server-edge/src/index.ts
@@ -1,0 +1,25 @@
+/**
+ * This is the internal API reference for the common code shared between
+ * LaunchDarkly Edge SDKs.
+ *
+ * Do not use this package directly in a cloud worker.
+ *
+ */
+import type { Info } from '@launchdarkly/js-server-sdk-common';
+import { LDClient } from './api';
+import validateOptions, { LDOptions, LDOptionsInternal } from './utils/validateOptions';
+
+export * from '@launchdarkly/js-server-sdk-common';
+export type { LDClient, LDOptions };
+
+/**
+ * Do not use this function directly.
+ *
+ * This is an internal API to be used directly only by LaunchDarkly Edge SDKs.
+ */
+export const init = (sdkKey: string, platformInfo: Info, options: LDOptionsInternal) => {
+  // this throws if options are invalid
+  validateOptions(sdkKey, options);
+
+  return new LDClient(sdkKey, platformInfo, options);
+};